Upate root tracking in URLTable
Entires in URL table now have the root used at the time stored in the
entry. This is use in static file processing to resolve inputs

This is needed as the internal table of a URL Table can be used in other
url tables. This provide a single output mapping, but allows different
roots to be used,

Loading templates outside of the current root system need this.

Loading templates has been updated to link new tables if the root of the
current template and the requested template are different.

Inheriting now takes an optional root argument to reset the root for any
resources it needs from its own hierrachy